requirements.txt
langchain-groq==0.2.1
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
sentence-transformers==3.3.1
cohere==5.13.4
redis==5.2.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
python-docx==1.1.2
PyPDF2==3.0.1
python-pptx==1.0.2
openpyxl==3.1.5
pydantic==2.10.4
gtts==2.5.4
google-generativeai==0.8.3
Pillow==10.4.0
ultralytics==8.3.235
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless>=4.6.0
